requirements.txt
numpy
opencv-python
pillow
pynput
pywin32
mss
pyserial
